drivers: stepper: refactor gpio_stepper_driver & add stepper_control
Refactor motion control related function from gpio_stepper_controller to
zephy stepper control driver
