+++ drivers/CMakeLists.txt
@@ -85,6 +85,7 @@ add_subdirectory_ifdef(CONFIG_SERIAL serial)
 add_subdirectory_ifdef(CONFIG_SMBUS smbus)
 add_subdirectory_ifdef(CONFIG_SPI spi)
 add_subdirectory_ifdef(CONFIG_STEPPER stepper)
+add_subdirectory_ifdef(CONFIG_STEPPER_CONTROL stepper_control)
 add_subdirectory_ifdef(CONFIG_SYSCON syscon)
 add_subdirectory_ifdef(CONFIG_SYS_CLOCK_EXISTS timer)
 add_subdirectory_ifdef(CONFIG_TEE tee)
