+++ drivers/Kconfig
@@ -83,6 +83,7 @@ source "drivers/sip_svc/Kconfig"
 source "drivers/smbus/Kconfig"
 source "drivers/spi/Kconfig"
 source "drivers/stepper/Kconfig"
+source "drivers/stepper_control/Kconfig"
 source "drivers/syscon/Kconfig"
 source "drivers/timer/Kconfig"
 source "drivers/usb/Kconfig"
