+++ drivers/stepper_control/CMakeLists.txt
@@ -1,0 +1,13 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025 Jilay Sandeep Pandya
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_syscall_header(${ZEPHYR_BASE}/include/zephyr/drivers/stepper_control.h)
+
+# zephyr-keep-sorted-start
+
+# zephyr-keep-sorted-stop
+
+zephyr_library()
+zephyr_library_property(ALLOW_EMPTY TRUE)
+
+zephyr_library_sources_ifdef(CONFIG_ZEPHYR_STEPPER_CONTROL stepper_control.c)
