+++ samples/drivers/stepper/generic/boards/nucleo_g071rb.overlay
@@ -1,6 +1,7 @@
 / {
 	aliases {
 		stepper = &gpio_stepper;
+		stepper-control = &stepper_control;
 	};
 };
 
@@ -13,5 +14,14 @@
 			<&gpioc 7 GPIO_ACTIVE_HIGH>,	/* D9 */
 			<&gpiob 0 GPIO_ACTIVE_HIGH>,	/* D10 */
 			<&gpioa 7 GPIO_ACTIVE_HIGH>;	/* D11 */
+	};
+};
+
+/ {
+	stepper_control: stepper_control {
+		compatible = "zephyr,stepper-control";
+		status = "okay";
+		stepper = <&gpio_stepper>;
+		step-tick-ns = <1000000>;
 	};
 };
