+++ samples/drivers/stepper/generic/prj.conf
@@ -1,3 +1,3 @@
-CONFIG_STEPPER=y
+CONFIG_STEPPER_CONTROL=y
 CONFIG_LOG=y
 CONFIG_INPUT=y
