+++ samples/drivers/stepper/generic/src/main.c
@@ -5,6 +5,7 @@
 
 #include <zephyr/device.h>
 #include <zephyr/drivers/gpio.h>
+#include <zephyr/drivers/stepper_control.h>
 #include <zephyr/drivers/stepper.h>
 #include <zephyr/input/input.h>
 #include <zephyr/kernel.h>
@@ -13,6 +14,7 @@
 LOG_MODULE_REGISTER(stepper, CONFIG_STEPPER_LOG_LEVEL);
 
 static const struct device *stepper = DEVICE_DT_GET(DT_ALIAS(stepper));
+static const struct device *stepper_control = DEVICE_DT_GET(DT_ALIAS(stepper_control));
 
 enum stepper_mode {
 	STEPPER_MODE_ENABLE,
@@ -31,11 +33,11 @@ static int32_t ping_pong_target_position =
 
 static K_SEM_DEFINE(stepper_generic_sem, 0, 1);
 
-static void stepper_callback(const struct device *dev, const enum stepper_event event,
-			     void *user_data)
+static void stepper_control_callback(const struct device *dev,
+				     const enum stepper_control_event event, void *user_data)
 {
 	switch (event) {
-	case STEPPER_EVENT_STEPS_COMPLETED:
+	case STEPPER_CONTROL_EVENT_STEPS_COMPLETED:
 		k_sem_give(&stepper_generic_sem);
 		break;
 	default:
@@ -71,9 +73,14 @@ int main(void)
 	}
 	LOG_DBG("stepper is %p, name is %s\n", stepper, stepper->name);
 
-	stepper_set_event_callback(stepper, stepper_callback, NULL);
-	stepper_set_reference_position(stepper, 0);
-	stepper_set_microstep_interval(stepper, CONFIG_STEP_INTERVAL_NS);
+	if (!device_is_ready(stepper_control)) {
+		LOG_ERR("Device %s is not ready\n", stepper_control->name);
+		return -ENODEV;
+	}
+
+	stepper_control_set_event_callback(stepper_control, stepper_control_callback, NULL);
+	stepper_control_set_reference_position(stepper_control, 0);
+	stepper_control_set_step_interval(stepper_control, CONFIG_STEP_INTERVAL_NS);
 
 	for (;;) {
 		k_sem_take(&stepper_generic_sem, K_FOREVER);
@@ -83,25 +90,25 @@ int main(void)
 			LOG_INF("mode: enable\n");
 			break;
 		case STEPPER_MODE_STOP:
-			stepper_stop(stepper);
+			stepper_control_stop(stepper_control);
 			LOG_INF("mode: stop\n");
 			break;
 		case STEPPER_MODE_ROTATE_CW:
-			stepper_run(stepper, STEPPER_DIRECTION_POSITIVE);
+			stepper_control_run(stepper_control, STEPPER_DIRECTION_POSITIVE);
 			LOG_INF("mode: rotate cw\n");
 			break;
 		case STEPPER_MODE_ROTATE_CCW:
-			stepper_run(stepper, STEPPER_DIRECTION_NEGATIVE);
+			stepper_control_run(stepper_control, STEPPER_DIRECTION_NEGATIVE);
 			LOG_INF("mode: rotate ccw\n");
 			break;
 		case STEPPER_MODE_PING_PONG_RELATIVE:
 			ping_pong_target_position *= -1;
-			stepper_move_by(stepper, ping_pong_target_position);
+			stepper_control_move_by(stepper_control, ping_pong_target_position);
 			LOG_INF("mode: ping pong relative\n");
 			break;
 		case STEPPER_MODE_PING_PONG_ABSOLUTE:
 			ping_pong_target_position *= -1;
-			stepper_move_to(stepper, ping_pong_target_position);
+			stepper_control_move_to(stepper_control, ping_pong_target_position);
 			LOG_INF("mode: ping pong absolute\n");
 			break;
 		case STEPPER_MODE_DISABLE:
@@ -118,7 +125,7 @@ static void monitor_thread(void)
 	for (;;) {
 		int32_t actual_position;
 
-		stepper_get_actual_position(stepper, &actual_position);
+		stepper_control_get_actual_position(stepper_control, &actual_position);
 		LOG_DBG("Actual position: %d\n", actual_position);
 		k_sleep(K_MSEC(CONFIG_MONITOR_THREAD_TIMEOUT_MS));
 	}
